+++ datanode/networkhistory/initialise_test.go
@@ -7,7 +7,6 @@ import (
 
 	"code.vegaprotocol.io/vega/datanode/networkhistory"
 	"code.vegaprotocol.io/vega/datanode/networkhistory/mocks"
-	"code.vegaprotocol.io/vega/datanode/networkhistory/store"
 	"code.vegaprotocol.io/vega/datanode/sqlstore"
 	"code.vegaprotocol.io/vega/logging"
 
@@ -16,6 +15,29 @@ import (
 	v2 "code.vegaprotocol.io/vega/protos/data-node/api/v2"
 	"github.com/stretchr/testify/assert"
 )
+
+type TestSegment struct {
+	HeightFrom               int64
+	HeightTo                 int64
+	PreviousHistorySegmentId string
+	HistorySegmentId         string
+}
+
+func (m TestSegment) GetPreviousHistorySegmentId() string {
+	return m.PreviousHistorySegmentId
+}
+
+func (m TestSegment) GetHistorySegmentId() string {
+	return m.HistorySegmentId
+}
+
+func (m TestSegment) GetFromHeight() int64 {
+	return m.HeightFrom
+}
+
+func (m TestSegment) GetToHeight() int64 {
+	return m.HeightTo
+}
 
 func TestInitialiseEmptyDataNode(t *testing.T) {
 	log := logging.NewTestLogger()
@@ -33,7 +55,6 @@ func TestInitialiseEmptyDataNode(t *testing.T) {
 			Segment: &v2.HistorySegment{
 				FromHeight:               1001,
 				ToHeight:                 2000,
-				ChainId:                  "testchainid",
 				HistorySegmentId:         "segment2",
 				PreviousHistorySegmentId: "segment1",
 			},
@@ -45,30 +66,28 @@ func TestInitialiseEmptyDataNode(t *testing.T) {
 	service.EXPECT().GetMostRecentHistorySegmentFromPeers(gomock.Any(), []int{}).Times(1).
 		Return(peerResponse, map[string]*v2.GetMostRecentNetworkHistorySegmentResponse{"peer1": peerResponse.Response}, nil)
 
-	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment2").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               1001,
-			HeightTo:                 2001,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "segment1",
-		},
-		HistorySegmentID: "segment2",
-	}, nil)
+	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment2").Times(1).Return(
+		TestSegment{HeightFrom: 1001, HeightTo: 2000, PreviousHistorySegmentId: "segment1", HistorySegmentId: "segment2"}, nil)
 
-	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               0,
-			HeightTo:                 1000,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "",
-		},
-		HistorySegmentID: "segment1",
-	}, nil)
+	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(
+		TestSegment{HeightFrom: 0, HeightTo: 1000, PreviousHistorySegmentId: "", HistorySegmentId: "segment1"}, nil)
 
 	gomock.InOrder(first, second)
 
-	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
 	service.EXPECT().GetDatanodeBlockSpan(gomock.Any()).Times(1).Return(sqlstore.DatanodeBlockSpan{}, nil)
+	service.EXPECT().ListAllHistorySegments().Times(1).Return([]networkhistory.Segment{
+		TestSegment{HeightFrom: 0, HeightTo: 1000},
+		TestSegment{HeightFrom: 1001, HeightTo: 2000},
+	}, nil)
+
+	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), networkhistory.ContiguousHistory{
+		HeightFrom: 0,
+		HeightTo:   2000,
+		SegmentsOldestFirst: []networkhistory.Segment{
+			TestSegment{HeightFrom: 0, HeightTo: 1000},
+			TestSegment{HeightFrom: 1001, HeightTo: 2000},
+		},
+	}, gomock.Any(), false).Times(1)
 
 	networkhistory.InitialiseDatanodeFromNetworkHistory(ctx, cfg, log, sqlstore.NewDefaultConfig().ConnectionConfig, service, []int{})
 }
@@ -89,7 +108,6 @@ func TestInitialiseNonEmptyDataNode(t *testing.T) {
 			Segment: &v2.HistorySegment{
 				FromHeight:               3001,
 				ToHeight:                 4000,
-				ChainId:                  "testchainid",
 				HistorySegmentId:         "segment4",
 				PreviousHistorySegmentId: "segment3",
 			},
@@ -101,29 +119,28 @@ func TestInitialiseNonEmptyDataNode(t *testing.T) {
 	service.EXPECT().GetMostRecentHistorySegmentFromPeers(gomock.Any(), []int{}).Times(1).
 		Return(peerResponse, map[string]*v2.GetMostRecentNetworkHistorySegmentResponse{"peer1": peerResponse.Response}, nil)
 
-	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment4").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               3001,
-			HeightTo:                 4000,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "segment3",
-		},
-		HistorySegmentID: "segment4",
-	}, nil)
+	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment4").Times(1).Return(
+		TestSegment{HeightFrom: 3001, HeightTo: 4000, PreviousHistorySegmentId: "segment3", HistorySegmentId: "segment4"}, nil)
 
-	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment3").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               2001,
-			HeightTo:                 3000,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "segment2",
-		},
-		HistorySegmentID: "segment3",
-	}, nil)
+	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment3").Times(1).Return(
+		TestSegment{HeightFrom: 2001, HeightTo: 3000, PreviousHistorySegmentId: "segment2", HistorySegmentId: "segment3"}, nil)
 
 	gomock.InOrder(first, second)
 
-	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
+	service.EXPECT().ListAllHistorySegments().Times(1).Return(
+		[]networkhistory.Segment{
+			TestSegment{HeightFrom: 2001, HeightTo: 3000},
+			TestSegment{HeightFrom: 3001, HeightTo: 4000},
+		}, nil)
+
+	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), networkhistory.ContiguousHistory{
+		HeightFrom: 2001,
+		HeightTo:   4000,
+		SegmentsOldestFirst: []networkhistory.Segment{
+			TestSegment{HeightFrom: 2001, HeightTo: 3000},
+			TestSegment{HeightFrom: 3001, HeightTo: 4000},
+		},
+	}, gomock.Any(), true).Times(1)
 
 	service.EXPECT().GetDatanodeBlockSpan(gomock.Any()).Times(1).Return(sqlstore.DatanodeBlockSpan{
 		FromHeight: 0,
@@ -150,7 +167,6 @@ func TestLoadingHistoryWithinDatanodeCurrentSpanDoesNothing(t *testing.T) {
 			Segment: &v2.HistorySegment{
 				FromHeight:               3001,
 				ToHeight:                 4000,
-				ChainId:                  "testchainid",
 				HistorySegmentId:         "segment4",
 				PreviousHistorySegmentId: "segment3",
 			},
@@ -187,7 +203,6 @@ func TestWhenMinimumBlockCountExceedsAvailableHistory(t *testing.T) {
 			Segment: &v2.HistorySegment{
 				FromHeight:               1001,
 				ToHeight:                 2000,
-				ChainId:                  "testchainid",
 				HistorySegmentId:         "segment2",
 				PreviousHistorySegmentId: "segment1",
 			},
@@ -198,30 +213,30 @@ func TestWhenMinimumBlockCountExceedsAvailableHistory(t *testing.T) {
 	service.EXPECT().GetMostRecentHistorySegmentFromPeers(gomock.Any(), []int{}).Times(1).
 		Return(peerResponse, map[string]*v2.GetMostRecentNetworkHistorySegmentResponse{"peer1": peerResponse.Response}, nil)
 
-	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment2").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               1001,
-			HeightTo:                 2001,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "segment1",
-		},
-		HistorySegmentID: "segment2",
-	}, nil)
+	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment2").Times(1).Return(
+		TestSegment{HeightFrom: 1001, HeightTo: 2000, PreviousHistorySegmentId: "segment1", HistorySegmentId: "segment2"}, nil)
 
-	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               0,
-			HeightTo:                 1000,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "",
-		},
-		HistorySegmentID: "segment1",
-	}, nil)
+	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(
+		TestSegment{HeightFrom: 0, HeightTo: 1000, PreviousHistorySegmentId: "", HistorySegmentId: "segment1"}, nil)
 
 	gomock.InOrder(first, second)
 
-	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
 	service.EXPECT().GetDatanodeBlockSpan(gomock.Any()).Times(1).Return(sqlstore.DatanodeBlockSpan{}, nil)
+
+	service.EXPECT().ListAllHistorySegments().Times(1).Return(
+		[]networkhistory.Segment{
+			TestSegment{HeightFrom: 0, HeightTo: 1000},
+			TestSegment{HeightFrom: 1001, HeightTo: 2000},
+		}, nil)
+
+	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), networkhistory.ContiguousHistory{
+		HeightFrom: 0,
+		HeightTo:   2000,
+		SegmentsOldestFirst: []networkhistory.Segment{
+			TestSegment{HeightFrom: 0, HeightTo: 1000},
+			TestSegment{HeightFrom: 1001, HeightTo: 2000},
+		},
+	}, gomock.Any(), false).Times(1)
 
 	networkhistory.InitialiseDatanodeFromNetworkHistory(ctx, cfg, log, sqlstore.NewDefaultConfig().ConnectionConfig,
 		service, []int{})
@@ -238,18 +253,23 @@ func TestInitialiseToASpecifiedSegment(t *testing.T) {
 	service := mocks.NewMockNetworkHistory(ctrl)
 	ctx := context.Background()
 
-	service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(store.SegmentIndexEntry{
-		SegmentMetaData: store.SegmentMetaData{
-			HeightFrom:               0,
-			HeightTo:                 1000,
-			ChainID:                  "testchainid",
-			PreviousHistorySegmentID: "",
-		},
-		HistorySegmentID: "segment1",
-	}, nil)
+	service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(
+		TestSegment{HeightFrom: 0, HeightTo: 1000, PreviousHistorySegmentId: "", HistorySegmentId: "segment1"}, nil)
 
-	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
 	service.EXPECT().GetDatanodeBlockSpan(gomock.Any()).Times(1).Return(sqlstore.DatanodeBlockSpan{}, nil)
+
+	service.EXPECT().ListAllHistorySegments().Times(1).Return(
+		[]networkhistory.Segment{
+			TestSegment{HeightFrom: 0, HeightTo: 1000},
+		}, nil)
+
+	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(), networkhistory.ContiguousHistory{
+		HeightFrom: 0,
+		HeightTo:   1000,
+		SegmentsOldestFirst: []networkhistory.Segment{
+			TestSegment{HeightFrom: 0, HeightTo: 1000},
+		},
+	}, gomock.Any(), false).Times(1)
 
 	networkhistory.InitialiseDatanodeFromNetworkHistory(ctx, cfg, log, sqlstore.NewDefaultConfig().ConnectionConfig,
 		service, []int{})
@@ -291,6 +311,62 @@ func TestAutoInitialiseWhenNoHistoryAvailableFromPeers(t *testing.T) {
 
 	assert.NotNil(t, networkhistory.InitialiseDatanodeFromNetworkHistory(ctx, cfg, log, sqlstore.NewDefaultConfig().ConnectionConfig,
 		service, []int{}))
+}
+
+func TestInitialiseEmptyDataNodeWhenMultipleContiguousHistories(t *testing.T) {
+	log := logging.NewTestLogger()
+	cfg := networkhistory.NewDefaultInitializationConfig()
+
+	ctrl := gomock.NewController(t)
+	service := mocks.NewMockNetworkHistory(ctrl)
+	ctx := context.Background()
+
+	peerResponse := &networkhistory.PeerResponse{
+		PeerAddr: "",
+		Response: &v2.GetMostRecentNetworkHistorySegmentResponse{
+			Segment: &v2.HistorySegment{
+				FromHeight:               6001,
+				ToHeight:                 7000,
+				HistorySegmentId:         "segment2",
+				PreviousHistorySegmentId: "segment1",
+			},
+			SwarmKey: "",
+		},
+	}
+
+	cfg.MinimumBlockCount = 1500
+	service.EXPECT().GetMostRecentHistorySegmentFromPeers(gomock.Any(), []int{}).Times(1).
+		Return(peerResponse, map[string]*v2.GetMostRecentNetworkHistorySegmentResponse{"peer1": peerResponse.Response}, nil)
+
+	first := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment2").Times(1).Return(
+		TestSegment{HeightFrom: 6001, HeightTo: 7000, PreviousHistorySegmentId: "segment1", HistorySegmentId: "segment2"}, nil)
+
+	second := service.EXPECT().FetchHistorySegment(gomock.Any(), "segment1").Times(1).Return(
+		TestSegment{HeightFrom: 5001, HeightTo: 6000, PreviousHistorySegmentId: "", HistorySegmentId: "segment1"}, nil)
+
+	gomock.InOrder(first, second)
+
+	service.EXPECT().GetDatanodeBlockSpan(gomock.Any()).Times(1).Return(sqlstore.DatanodeBlockSpan{}, nil)
+	service.EXPECT().ListAllHistorySegments().Times(1).Return(
+		[]networkhistory.Segment{
+			TestSegment{HeightFrom: 2001, HeightTo: 3000},
+			TestSegment{HeightFrom: 3001, HeightTo: 4000},
+
+			TestSegment{HeightFrom: 5001, HeightTo: 6000},
+			TestSegment{HeightFrom: 6001, HeightTo: 7000},
+		},
+		nil)
+	service.EXPECT().LoadNetworkHistoryIntoDatanode(gomock.Any(),
+		networkhistory.ContiguousHistory{
+			HeightFrom: 5001,
+			HeightTo:   7000,
+			SegmentsOldestFirst: []networkhistory.Segment{
+				TestSegment{HeightFrom: 5001, HeightTo: 6000},
+				TestSegment{HeightFrom: 6001, HeightTo: 7000},
+			},
+		}, gomock.Any(), false).Times(1)
+
+	networkhistory.InitialiseDatanodeFromNetworkHistory(ctx, cfg, log, sqlstore.NewDefaultConfig().ConnectionConfig, service, []int{})
 }
 
 func TestSelectRootSegment(t *testing.T) {